scope GlobMap to ruletable package
